[VOID] ViewerBuffer: Split core functionality of the buffer into PlayerBuffer
ViewerBuffer inherits PlayerBuffer and adds cache functionality to it which is used by the Player
