+++ src/VoidUi/Player/ViewerBuffer.cpp
@@ -22,22 +22,39 @@ static const std::size_t s_MaxBuffers = 1000;
 //     return stream;
 // }
 
+/// Tasks
+
+class CacheNextFrameTask : public QRunnable
+{
+public:
+    CacheNextFrameTask(ViewerBuffer* parent) : m_Parent(parent) {}
+    inline void run() override { m_Parent->CacheNextFrame(); }
+
+private:
+    ViewerBuffer* m_Parent;
+};
+
+class CachePreviousFrameTask : public QRunnable
+{
+public:
+    CachePreviousFrameTask(ViewerBuffer* parent) : m_Parent(parent) {}
+    inline void run() override { m_Parent->CachePreviousFrame(); }
+
+private:
+    ViewerBuffer* m_Parent;
+};
+
+/// ViewerBuffer
+
 ViewerBuffer::ViewerBuffer(QObject* parent)
-    : QObject(parent)
+    : PlayerBuffer(parent)
     , m_Name("Viewer")
     , m_Color(130, 110, 190)    // Purple
     , m_Player(nullptr)
-    , m_Clip(std::make_shared<MediaClip>())
-    , m_Track(std::make_shared<PlaybackTrack>())
-    , m_Sequence(std::make_shared<PlaybackSequence>())
-    , m_Playlist(nullptr)
     , m_MaxMemory(VoidPreferences::Instance().GetCacheMemory() * 1024 * 1024 * 1024) // 1 GB by default
     , m_Framesize(1)
     , m_Capacity(10)
-    , m_Startframe(0)
-    , m_Endframe(1)
     , m_LastCached(0)
-    , m_PlayingComponent(PlayableComponent::Clip)
     , m_State(PlayState::Forwards)
     , m_BackBuffer(3)
     , m_Active(false)
@@ -45,112 +62,16 @@ ViewerBuffer::ViewerBuffer(QObject* parent)
     m_ThreadPool.setMaxThreadCount(VoidPreferences::Instance().GetCacheThreads());
     connect(&m_CacheTimer, &QTimer::timeout, this, &ViewerBuffer::Update, Qt::DirectConnection);
     connect(&VoidPreferences::Instance(), &VoidPreferences::updated, this, &ViewerBuffer::SettingsUpdated);
+    connect(this, &PlayerBuffer::mediaAboutToBeUpdated, this, &ViewerBuffer::Refresh, Qt::DirectConnection);
+    connect(this, &PlayerBuffer::mediaUpdated, this, &ViewerBuffer::ResetMedia);
+    connect(this, &PlayerBuffer::rangeUpdated, this, [this](v_frame_t start, v_frame_t end) -> void
+    {
+        m_BackBuffer = std::min(10, std::max(3, static_cast<int>(((m_Endframe - m_Startframe) + 1) * 0.02)));
+    }, Qt::DirectConnection);
 }
 
 ViewerBuffer::~ViewerBuffer()
 {
-}
-
-void ViewerBuffer::Set(const SharedMediaClip& media)
-{
-    Refresh();
-
-    m_Clip = media;
-    m_Clip->SetColor(m_Color);
-
-    m_PlayingComponent = PlayableComponent::Clip;
-    UpdateRange(m_Clip->FirstFrame(), m_Clip->LastFrame());
-    EnsureCached(media->FirstFrame());
-
-    emit playlistUpdated(nullptr);
-    CacheAvailable();
-}
-
-void ViewerBuffer::Set(const SharedPlaybackTrack& track)
-{
-    Refresh();
-
-    m_Track = track;
-    m_Track->SetColor(m_Color);
-
-    m_PlayingComponent = PlayableComponent::Track;
-    UpdateRange(m_Track->StartFrame(), m_Track->EndFrame());
-
-    if (auto item = m_Track->GetTrackItem(m_Startframe))
-        EnsureCached(item->StartFrame());
-
-    CacheAvailable();
-}
-
-void ViewerBuffer::Set(const SharedPlaybackSequence& sequence)
-{
-    Refresh();
-
-    m_Sequence = sequence;
-    m_PlayingComponent = PlayableComponent::Sequence;
-    UpdateRange(m_Sequence->StartFrame(), m_Sequence->EndFrame());
-
-    if (auto item = m_Sequence->GetTrackItem(m_Startframe))
-        EnsureCached(item->StartFrame());
-
-    CacheAvailable();
-}
-
-void ViewerBuffer::Set(const std::vector<SharedMediaClip>& media)
-{
-    Refresh();
-
-    m_Track->Clear();
-    for (const SharedMediaClip& media : media)
-        m_Track->AddMedia(media);
-
-    m_PlayingComponent = PlayableComponent::Track;
-    UpdateRange(m_Track->StartFrame(), m_Track->EndFrame());
-
-    if (auto item = m_Track->GetTrackItem(m_Startframe))
-        EnsureCached(item->StartFrame());
-
-    CacheAvailable();
-}
-
-void ViewerBuffer::SetGrid(Playlist* playlist)
-{
-    Refresh();
-    if (m_Playlist)
-        disconnect(m_Playlist, &Playlist::updated, this, &ViewerBuffer::updated);
-
-    m_Playlist = playlist;
-    m_Clip = playlist->CurrentMedia();
-
-    connect(m_Playlist, &Playlist::updated, this, &ViewerBuffer::updated);
-
-    m_PlayingComponent = PlayableComponent::Grid;
-    UpdateRange(m_Clip->FirstFrame(), m_Clip->LastFrame());
-
-    emit playlistUpdated(playlist);
-
-    EnsureCached(m_Clip->FirstFrame());
-    CacheAvailable();
-}
-
-void ViewerBuffer::SetPlaylist(Playlist* playlist)
-{
-    Refresh();
-    if (m_Playlist)
-        disconnect(m_Playlist, &Playlist::updated, this, &ViewerBuffer::updated);
-
-    m_Playlist = playlist;
-    m_Clip = playlist->CurrentMedia();
-
-    connect(m_Playlist, &Playlist::updated, this, &ViewerBuffer::updated);
-
-    m_PlayingComponent = PlayableComponent::Playlist;
-    UpdateRange(m_Clip->FirstFrame(), m_Clip->LastFrame());
-
-    emit playlistUpdated(playlist);
-
-    EnsureCached(m_Clip->FirstFrame());
-    CacheAvailable();
 }
 
 void ViewerBuffer::SetColor(const QColor& color)
@@ -161,41 +82,6 @@ void ViewerBuffer::SetColor(const QColor& color)
     /* Update entities with the color change */
     m_Clip->SetColor(color);
     m_Track->SetColor(color);
-}
-
-SharedPlaybackTrack ViewerBuffer::ActiveTrack() const
-{
-    switch (m_PlayingComponent)
-    {
-        case PlayableComponent::Clip:
-            /**
-             * A clip is a separate entity and does not relate to track
-             * If a clip is active then it should be the one currently playing and no track
-             */
-            return nullptr;
-        case PlayableComponent::Track:
-            /**
-             * If the current playing component itself is a track, then it will be considered
-             * as the active track, this could involve in adding elements to the active track
-             */
-            return m_Track;
-        case PlayableComponent::Sequence:
-            /**
-             * If a sequence is currently playing, then this returns the currently active track
-             * Meaning the track which is on the very top and also which is both enabled and visible
-             * for it to recieve new entities and also show entities
-             */
-            /**
-             * TODO: Need to think about, if this is really needed when right clicking > add media to sequence?
-             * if added, where does the media go to? at the last of track or clears it?
-             */
-            return m_Sequence->ActiveVideoTrack();
-        case PlayableComponent::Grid:
-        case PlayableComponent::Playlist:
-            return nullptr;
-    }
-
-    return nullptr;
 }
 
 void ViewerBuffer::Refresh()
@@ -227,151 +113,10 @@ const FloatImage ViewerBuffer::Image(const v_frame_t frame)
     return nullptr;
 }
 
-BufferData ViewerBuffer::MData(const v_frame_t frame, bool nearest)
-{
-    BufferData d;
-    switch (m_PlayingComponent)
-    {
-        case PlayableComponent::Track:
-            d.image = m_Track->Image(frame);
-            break;
-        case PlayableComponent::Sequence:
-            d.image = m_Sequence->Image(frame);
-            break;
-        default:
-            d.image = m_Clip->Image(frame);
-            d.annotation = m_Clip->Annotation(frame);
-    }
-
-    return d;
-}
-
-std::vector<FloatImage> ViewerBuffer::GridFrame(const v_frame_t frame)
-{
-    std::vector<FloatImage> grid;
-    grid.reserve(m_Playlist->Size());
-
-    if (m_PlayingComponent == PlayableComponent::Grid)
-    {
-        for (auto& media : m_Playlist->AllMedia())
-        {
-            FloatImage image = VOID_NAMESPACE::Image<float>::Create();
-            media->Image(media->Contains(frame) ? frame : media->NearestFrame(frame), image);
-            grid.push_back(image);
-        }
-    }
-
-    return grid;
-}
-
-SharedMediaClip ViewerBuffer::Media(const v_frame_t frame)
-{
-    switch (m_PlayingComponent)
-    {
-        case PlayableComponent::Sequence: return m_Sequence->Media(frame);
-        case PlayableComponent::Track: return m_Track->Media(frame);
-        case PlayableComponent::Grid:
-        case PlayableComponent::Playlist:
-        case PlayableComponent::Clip:
-        default: return m_Clip;
-    }
-}
-
-SharedTrackItem ViewerBuffer::TrackItem(const v_frame_t frame)
-{
-    return (m_PlayingComponent == PlayableComponent::Sequence)
-            ? m_Sequence->GetTrackItem(frame)
-            : m_Track->GetTrackItem(frame);
-}
-
 void ViewerBuffer::Clear()
 {
     ClearCache();
-    m_Clip = std::make_shared<MediaClip>();
-}
-
-bool ViewerBuffer::Playing(const SharedMediaClip& media) const
-{
-    // TODO: Check if we may have a use case around if the clip being compared against be a part of sequence or track as well
-    return m_PlayingComponent == PlayableComponent::Clip ? media == m_Clip : false;
-}
-
-void ViewerBuffer::SetAnnotation(const v_frame_t frame, const Renderer::SharedAnnotation& annotation)
-{
-    /**
-     * For now Just adding the annotation on the Media Clip
-     * TODO: The Active sequence should also have a way to save Annotations to
-     * that way it's not applied on to the TrackItem but on the sequence itself (or maybe check behaviours around this)
-     */
-    if (m_PlayingComponent == PlayableComponent::Clip)
-        m_Clip->SetAnnotation(frame, annotation);
-}
-
-void ViewerBuffer::RemoveAnnotation(const v_frame_t frame)
-{
-    /**
-     * For now Just removing the annotation from the Media Clip
-     * TODO: The Active sequence should also have a way to save Annotations to and remove from
-     * that way it's not happening on to TrackItem but on the sequence itself (or maybe check behaviours around this)
-     */
-    if (m_PlayingComponent == PlayableComponent::Clip)
-        m_Clip->RemoveAnnotation(frame);
-}
-
-bool ViewerBuffer::NextMedia()
-{
-    if (m_PlayingComponent == PlayableComponent::Playlist)
-    {
-        Refresh();
-        m_Clip = m_Playlist->NextMedia();
-
-        UpdateRange(m_Clip->FirstFrame(), m_Clip->LastFrame());
-        EnsureCached(m_Clip->FirstFrame());
-        CacheAvailable();
-
-        return true;
-    }
-
-    return false;
-}
-
-bool ViewerBuffer::PreviousMedia()
-{
-    if (m_PlayingComponent == PlayableComponent::Playlist)
-    {
-        Refresh();
-        m_Clip = m_Playlist->PreviousMedia();
-
-        UpdateRange(m_Clip->FirstFrame(), m_Clip->LastFrame());
-        EnsureCached(m_Clip->FirstFrame());
-        CacheAvailable();
-
-        return true;
-    }
-
-    return false;
-}
-
-bool ViewerBuffer::ResetPlaylistMedia()
-{
-    /**
-     * This is supposed to be invoked when we have reset the current index of the playlist
-     * and we want that to be played now,
-     * this clears existing cache and fetches the current media info from the playlist if available
-     */
-    if (m_PlayingComponent == PlayableComponent::Playlist)
-    {
-        Refresh();
-        m_Clip = m_Playlist->CurrentMedia();
-
-        UpdateRange(m_Clip->FirstFrame(), m_Clip->LastFrame());
-        EnsureCached(m_Clip->FirstFrame());
-        CacheAvailable();
-
-        return true;
-    }
-
-    return false;
+    ResetMedia();
 }
 
 void ViewerBuffer::StartPlaybackCache(const PlayState& state)
@@ -441,53 +186,6 @@ void ViewerBuffer::Update()
     VOID_LOG_INFO("Update Cache.....");
 }
 
-void ViewerBuffer::UpdateRange(v_frame_t start, v_frame_t end)
-{
-    m_Startframe = start;
-    m_Endframe = end;
-
-    /**
-     * The Back buffer refers to the amount of frames which stay behind the playhead
-     * This needs to be between 3 - 10 depending on the overall size of the entity
-     * being played
-     */
-    m_BackBuffer = std::min(10, std::max(3, static_cast<int>(((m_Endframe - m_Startframe) + 1) * 0.02)));
-}
-
-bool ViewerBuffer::Request(v_frame_t frame, bool evict)
-{
-    if (Completed())
-    {
-        m_CacheTimer.stop();
-        return false;
-    }
-
-    if (Maxxed())
-    {
-        if (evict)
-        {
-            if (m_State == PlayState::Backwards)
-            {
-                EvictBack();
-                m_Numbers.push_front(frame);
-            }
-            else
-            {
-                EvictFront();
-                m_Numbers.push_back(frame);
-            }
-
-            return true;
-        }
-
-        // Cannot grant this request as we do not have enough memory
-        return false;
-    }
-
-    m_State == PlayState::Backwards ? m_Numbers.push_front(frame) : m_Numbers.push_back(frame);
-    return true;
-}
-
 void ViewerBuffer::SetFramesize(std::size_t size)
 {
     if (size == m_Framesize)
@@ -528,6 +226,65 @@ void ViewerBuffer::Cache(v_frame_t frame)
 
         Store(frame);
     }
+}
+
+bool ViewerBuffer::Request(v_frame_t frame, bool evict)
+{
+    if (Completed())
+    {
+        m_CacheTimer.stop();
+        return false;
+    }
+
+    if (Maxxed())
+    {
+        if (evict)
+        {
+            if (m_State == PlayState::Backwards)
+            {
+                EvictBack();
+                m_Numbers.push_front(frame);
+            }
+            else
+            {
+                EvictFront();
+                m_Numbers.push_back(frame);
+            }
+
+            return true;
+        }
+
+        // Cannot grant this request as we do not have enough memory
+        return false;
+    }
+
+    m_State == PlayState::Backwards ? m_Numbers.push_front(frame) : m_Numbers.push_back(frame);
+    return true;
+}
+
+bool ViewerBuffer::Completed() const
+{
+    /**
+     * The cache/buffer is said to be completed if the internal frame store has the start
+     * and the end frames present and the overall duration is same or exceeds the duration of playback
+     *
+     */
+    return m_Numbers.size() >= InternalDuration() && Buffered(InternalStartframe()) && Buffered(InternalEndframe());
+}
+
+void ViewerBuffer::EnsureCached(v_frame_t frame)
+{
+    if (Buffered(frame))
+        return;
+
+    VOID_LOG_INFO("Force Caching Frame -- : {0}", frame);
+    {
+        std::lock_guard<std::mutex> lock(m_Mutex);
+        m_LastCached = frame;
+    }
+
+    Request(frame, true);
+    Cache(frame);
 }
 
 void ViewerBuffer::EvictFront()
@@ -586,21 +343,6 @@ void ViewerBuffer::Store(v_frame_t frame)
     m_Player->AddCacheFrame(frame);
 }
 
-void ViewerBuffer::EnsureCached(v_frame_t frame)
-{
-    if (Buffered(frame))
-        return;
-
-    VOID_LOG_INFO("Force Caching Frame -- : {0}", frame);
-    {
-        std::lock_guard<std::mutex> lock(m_Mutex);
-        m_LastCached = frame;
-    }
-
-    Request(frame, true);
-    Cache(frame);
-}
-
 void ViewerBuffer::Recache()
 {
     ClearCache();
@@ -643,36 +385,15 @@ void ViewerBuffer::CacheAvailable()
     }
 }
 
-v_frame_t ViewerBuffer::InternalStartframe() const
+void ViewerBuffer::ResetMedia()
 {
-    return (m_Player) ? std::max(m_Startframe, m_Player->Startframe()) : m_Startframe;
-}
-
-v_frame_t ViewerBuffer::InternalEndframe() const
-{
-    return (m_Player) ? std::min(m_Endframe, m_Player->Endframe()) : m_Endframe;
-}
-
-int ViewerBuffer::InternalDuration() const
-{
-    // return InternalEndframe() - InternalStartframe() + 1;
-    return m_Endframe - m_Startframe + 1;
-}
-
-bool ViewerBuffer::Completed() const
-{
-    /**
-     * The cache/buffer is said to be completed if the internal frame store has the start
-     * and the end frames present
-     *
-     */
-    return m_Numbers.size() >= InternalDuration();
+    EnsureCached(m_Startframe);
+    CacheAvailable();
 }
 
 v_frame_t ViewerBuffer::GetNextFrame()
 {
     std::lock_guard<std::mutex> lock(m_Mutex);
-
     v_frame_t frame = 0;
 
     if (m_LastCached >= m_Endframe)
@@ -688,7 +409,6 @@ v_frame_t ViewerBuffer::GetNextFrame()
 v_frame_t ViewerBuffer::GetPreviousFrame()
 {
     std::lock_guard<std::mutex> lock(m_Mutex);
-
     v_frame_t frame = 0;
 
     if (m_LastCached <= m_Startframe)
